pyproject.toml
[project]
name = "wireguard-configure-generator-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "cryptography>=45.0.2",
    "jinja2>=3.1.6",
    "jsonschema>=4.23.0",
    "matplotlib>=3.10.3",
    "networkx>=3.4.2",
    "sqlalchemy>=2.0.41",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
]
